added class for token, removing user (added kernel, job), remove send column from db

+++ database/migrations/2018_10_17_173857_add_confirmation_column_to_users_table.php
@@ -4,7 +4,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class AddConfirmationAndSendColumnsToUsersTable extends Migration
+class AddConfirmationColumnToUsersTable extends Migration
 {
     /**
      * Run the migrations.
@@ -15,7 +15,6 @@ class AddConfirmationAndSendColumnsToUsersTable extends Migration
     {
         Schema::table('users', function(Blueprint $table) {
             $table->boolean('confirmation')->nullable();
-            $table->timestamp('send')->nullable();
         });
     }
 
@@ -27,10 +26,7 @@ class AddConfirmationAndSendColumnsToUsersTable extends Migration
     public function down()
     {
         Schema::table('users', function(Blueprint $table) {
-            $table->dropColumn([
-                'confirmation',
-                'send',
-            ]);
+            $table->dropColumn(['confirmation']);
         });
     }
 }
